+++ src/blackmagic_io/__init__.py
@@ -459,7 +459,7 @@ except ImportError:
     # C++ extension not built yet
     pass
 
-__version__ = "0.17.0b2"
+__version__ = "0.17.0b3"
 __author__ = "Nick Shaw"
 
 __all__ = [